requirements.txt
flask==3.1.3
pika==1.3.2
prometheus_client==0.24.1
logging_json==0.6.0
pybreaker==1.2.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-flask==0.51b0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation-wsgi==0.51b0


